+++ src/platforms/pc/thread_with_pthread.c
@@ -402,7 +402,7 @@ tryagain:
   }
 #else  /*HAVE_SEM_TIMEDWAIT*/
 
-  uint32_t start = time_now_ms();
+  uint64_t start = time_now_ms();
   return_value_if_fail(semaphore != NULL, RET_BAD_PARAMS);
 
   do {
